Move dictionary and number to sys crate
